updated facade, service and queryhandlers to not use productdto. moved folders around ugly bug fix for product service as i couldn't figure out bean issue.
