--- src/main/kotlin/org/ptss/support/api/dtos/ProductDto.kt
@@ -1,8 +1,0 @@
-package org.ptss.support.api.dtos
-
-data class ProductDto(
-    val name: String,
-    val description: String,
-    val media: List<String>
-)
-
