+++ src/main/kotlin/org/ptss/support/core/facades/ProductFacade.kt
@@ -1,16 +1,15 @@
 package org.ptss.support.core.facades
 
 import jakarta.enterprise.context.ApplicationScoped
+import jakarta.inject.Inject
 import org.ptss.support.api.dtos.requests.CreateProductRequest
 import org.ptss.support.api.dtos.responses.ProductResponse
 import org.ptss.support.core.services.ProductService
-import org.ptss.support.core.context.RequestContext
 import org.ptss.support.core.mappers.ProductMapper
 
 @ApplicationScoped
-class ProductFacade(
-    private val productService: ProductService,
-    private val requestContext: RequestContext
+class ProductFacade @Inject constructor(
+    private val productService: ProductService
 ) {
     suspend fun getAllProducts(): List<ProductResponse> =
         productService.getAllProductsAsync()
