+++ src/main/kotlin/org/ptss/support/core/services/ProductService.kt
@@ -2,22 +2,22 @@ package org.ptss.support.core.services
 
 import org.ptss.support.domain.commands.CreateProductCommand
 import org.ptss.support.domain.interfaces.cqrs.ICommandHandler
-import org.ptss.support.domain.interfaces.cqrs.IQueryHandler
+import org.ptss.support.domain.interfaces.queries.IQueryHandler
 import org.ptss.support.domain.querries.GetAllProductsQuery
 import org.ptss.support.domain.querries.GetProductByIdQuery
 import org.ptss.support.infrastructure.util.executeWithExceptionLoggingAsync
 import jakarta.enterprise.context.ApplicationScoped
 import org.ptss.support.domain.models.Product
-import org.slf4j.Logger
+import org.ptss.support.infrastructure.handlers.queries.product.GetAllProductsQueryHandler
 import org.slf4j.LoggerFactory
 
 @ApplicationScoped
 class ProductService(
     private val getProductByIdHandler: IQueryHandler<GetProductByIdQuery, Product?>,
-    private val getAllProductsHandler: IQueryHandler<GetAllProductsQuery, List<Product>>,
-    private val createProductHandler: ICommandHandler<CreateProductCommand, String>,
-    private val logger: Logger = LoggerFactory.getLogger(ProductService::class.java)
+    private val getAllProductsHandler: GetAllProductsQueryHandler, //used to be iqueryhandler, but bean error. I don't know why :'( pls help
+    private val createProductHandler: ICommandHandler<CreateProductCommand, String>
 ) {
+    private val logger = LoggerFactory.getLogger(ProductService::class.java)
     suspend fun getProductByIdAsync(productId: String): Product? {
         return logger.executeWithExceptionLoggingAsync(
             operation = { getProductByIdHandler.handleAsync(GetProductByIdQuery(productId)) },
