+++ src/main/kotlin/org/ptss/support/domain/interfaces/queries/IQueryHandler.kt
@@ -1,4 +1,4 @@
-package org.ptss.support.domain.interfaces.cqrs
+package org.ptss.support.domain.interfaces.queries
 
 interface IQueryHandler<in TQuery, TResult> {
     suspend fun handleAsync(query: TQuery): TResult
